# docs/api/openapi-spec.yaml
openapi: 3.0.3
info:
  title: CRM Platform API
  description: |
    Comprehensive API documentation for the CRM Platform microservices including
    authentication, user management, and security features.
    
    ## Authentication
    
    All protected endpoints require a valid JWT token in the Authorization header:
    ```
    Authorization: Bearer <access_token>
    ```
    
    ## Multi-Tenancy
    
    Most endpoints require a tenant context header:
    ```
    X-Tenant-ID: <tenant_uuid>
    ```
    
    ## Rate Limiting
    
    API endpoints are rate limited. Check response headers for current limits:
    - `X-RateLimit-Limit`: Request limit per window
    - `X-RateLimit-Remaining`: Remaining requests in current window
    - `X-RateLimit-Reset`: Window reset time (Unix timestamp)
    
  version: 1.0.0
  contact:
    name: CRM Platform API Support
    email: api-support@example.com
  license:
    name: MIT
    url: https://opensource.org/licenses/MIT
servers:
  - url: http://localhost:8081/api/v1
    description: Authentication Service (Development)
  - url: http://localhost:8082/api/v1
    description: User Management Service (Development)
  - url: http://localhost:8083/api/v1
    description: Tenant Service (Development)
  - url: https://api.crm-platform.com/v1
    description: Production Environment
security:
  - BearerAuth: []
  - TenantContext: []
paths:
...
  /users:
    get:
      tags:
        - User Management
      summary: List Users
      description: Retrieve a paginated list of users
      operationId: listUsers
      parameters:
        - $ref: '#/components/parameters/PageParam'
        - $ref: '#/components/parameters/SizeParam'
        - $ref: '#/components/parameters/SortParam'
        - name: status
          in: query
          schema:
            $ref: '#/components/schemas/UserStatus'
        - name: role
          in: query
          schema:
            type: string
        - name: search
          in: query
          schema:
            type: string
      responses:
        '200':
          description: Users retrieved successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/UserListResponse'

    post:
      tags:
        - User Management
      summary: Create User
      description: Create a new user account
      operationId: createUser
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CreateUserRequest'
      responses:
        '201':
          description: User created successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/CreateUserResponse'
        '400':
          $ref: '#/components/responses/BadRequest'
        '409':
          $ref: '#/components/responses/Conflict'
  /users/{userId}:
    get:
      tags:
        - User Management
      summary: Get User by ID
      description: Retrieve detailed information about a specific user
      operationId: getUserById
      parameters:
        - $ref: '#/components/parameters/UserIdParam'
      responses:
        '200':
          description: User retrieved successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/UserDetailResponse'
        '404':
          $ref: '#/components/responses/NotFound'

    put:
      tags:
        - User Management
      summary: Update User
      description: Update user information
      operationId: updateUser
      parameters:
        - $ref: '#/components/parameters/UserIdParam'
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/UpdateUserRequest'
      responses:
        '200':
          description: User updated successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/SuccessResponse'
        '404':
          $ref: '#/components/responses/NotFound'

    delete:
      tags:
        - User Management
      summary: Delete User
      description: Delete a user account (soft delete by default)
      operationId: deleteUser
      parameters:
        - $ref: '#/components/parameters/UserIdParam'
        - name: hard_delete
          in: query
          schema:
            type: boolean
            default: false
        - name: transfer_data_to
          in: query
          schema:
            type: string
            format: uuid
      responses:
        '200':
          description: User deleted successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/SuccessResponse'
...
components:
  securitySchemes:
    BearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
      description: JWT access token for authentication
    TenantContext:
      type: apiKey
      in: header
      name: X-Tenant-ID
      description: Tenant context identifier
  parameters:
    PageParam:
      name: page
      in: query
      description: Page number (0-based)
      schema:
        type: integer
        minimum: 0
        default: 0
    SizeParam:
      name: size
      in: query
      description: Page size
      schema:
        type: integer
        minimum: 1
        maximum: 100
        default: 20
    SortParam:
      name: sort
      in: query
      description: Sort criteria (field,direction)
      schema:
        type: string
        default: "createdAt,desc"
    UserIdParam:
      name: userId
      in: path
      required: true
      description: User identifier
      schema:
        type: string
        format: uuid
  schemas:
...
    UserInfo:
      type: object
      properties:
        id:
          type: string
          format: uuid
        username:
          type: string
        email:
          type: string
          format: email
        first_name:
          type: string
        last_name:
          type: string
        phone_number:
          type: string
        status:
          $ref: '#/components/schemas/UserStatus'
        roles:
          type: array
          items:
            type: string
        tenant_id:
          type: string
          format: uuid
    UserStatus:
      type: string
      enum:
        - ACTIVE
        - INACTIVE
        - SUSPENDED
        - PENDING_VERIFICATION
    CreateUserRequest:
      type: object
      required:
        - username
        - email
        - first_name
        - last_name
      properties:
        username:
          type: string
          example: "jane.smith"
        email:
          type: string
          format: email
          example: "jane.smith@example.com"
        first_name:
          type: string
          example: "Jane"
        last_name:
          type: string
          example: "Smith"
        phone_number:
          type: string
          example: "+1234567891"
        job_title:
          type: string
          example: "Product Manager"
        department:
          type: string
          example: "Product"
        manager_id:
          type: string
          format: uuid
        timezone:
          type: string
          example: "America/Los_Angeles"
        language:
          type: string
          example: "en"
        roles:
          type: array
          items:
            type: string
        teams:
          type: array
          items:
            type: string
            format: uuid
        custom_fields:
          type: object
          additionalProperties: true
        send_welcome_email:
          type: boolean
          default: true
    CreateUserResponse:
      type: object
      properties:
        success:
          type: boolean
        data:
          type: object
          properties:
            id:
              type: string
              format: uuid
            username:
              type: string
            email:
              type: string
              format: email
            status:
              $ref: '#/components/schemas/UserStatus'
            created_at:
              type: string
              format: date-time
            temporary_password:
              type: string
              description: Temporary password for first login
            password_reset_required:
              type: boolean
    UpdateUserRequest:
      type: object
      properties:
        first_name:
          type: string
        last_name:
          type: string
        phone_number:
          type: string
        job_title:
          type: string
        department:
          type: string
        timezone:
          type: string
        custom_fields:
          type: object
          additionalProperties: true
    UserListResponse:
      type: object
      properties:
        success:
          type: boolean
        data:
          $ref: '#/components/schemas/PagedResponse'
    UserDetailResponse:
      type: object
      properties:
        success:
          type: boolean
        data:
          $ref: '#/components/schemas/UserDetail'
    UserDetail:
      allOf:
        - $ref: '#/components/schemas/UserInfo'
        - type: object
          properties:
            job_title:
              type: string
            department:
              type: string
            manager_id:
              type: string
              format: uuid
            email_verified:
              type: boolean
            phone_verified:
              type: boolean
            created_at:
              type: string
              format: date-time
            updated_at:
              type: string
              format: date-time
            last_login_at:
              type: string
              format: date-time
            timezone:
              type: string
            language:
              type: string
            profile_image_url:
              type: string
              format: uri
            roles:
              type: array
              items:
                $ref: '#/components/schemas/UserRole'
            teams:
              type: array
              items:
                $ref: '#/components/schemas/UserTeam'
            permissions:
              type: array
              items:
                type: string
            custom_fields:
              type: object
              additionalProperties: true
    UserRole:
      type: object
      properties:
        id:
          type: string
          format: uuid
        name:
          type: string
        description:
          type: string
        assigned_at:
          type: string
          format: date-time
    UserTeam:
      type: object
      properties:
        id:
          type: string
          format: uuid
        name:
          type: string
        role:
          type: string
          enum: [MEMBER, LEAD, MANAGER]
        joined_at:
          type: string
          format: date-time
...
    SuccessResponse:
      type: object
      properties:
        success:
          type: boolean
          example: true
        data:
          type: object
          properties:
            message:
              type: string
              example: "Operation completed successfully"
    ErrorResponse:
      type: object
      properties:
        success:
          type: boolean
          example: false
        error:
          type: object
          properties:
            code:
              type: string
              example: "INVALID_CREDENTIALS"
            message:
              type: string
              example: "Invalid username or password"
            details:
              type: string
              example: "Additional error details"
        meta:
          $ref: '#/components/schemas/ResponseMeta'
    PagedResponse:
      type: object
      properties:
        content:
          type: array
          items:
            type: object
        page:
          type: integer
          example: 0
        size:
          type: integer
          example: 20
        total_elements:
          type: integer
          example: 150
        total_pages:
          type: integer
          example: 8
    ResponseMeta:
      type: object
      properties:
        timestamp:
          type: string
          format: date-time
        request_id:
          type: string
          format: uuid
        pagination:
          type: object
          properties:
            page:
              type: integer
            size:
              type: integer
            total:
              type: integer
  responses:
    BadRequest:
      description: Bad Request
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
          example:
            success: false
            error:
              code: "INVALID_REQUEST"
              message: "Invalid request format"
              details: "Required field 'email' is missing"
...
    NotFound:
      description: Not Found
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
          example:
            success: false
            error:
              code: "RESOURCE_NOT_FOUND"
              message: "Requested resource not found"
    Conflict:
      description: Conflict
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
          example:
            success: false
            error:
              code: "RESOURCE_CONFLICT"
              message: "Resource already exists"
...